+++ app/src/main/java/com/example/studenttaskmanagement/model/TaskNotification.java
@@ -1,34 +1,34 @@
 package com.example.studenttaskmanagement.model;
 
 public class TaskNotification {
-    private int id;
-    private int taskId;
+    private long id;
+    private long taskId;
     private String notifyTime;
     private int isSent;
 
     public TaskNotification() {
     }
 
-    public TaskNotification(int id, int taskId, String notifyTime, int isSent) {
+    public TaskNotification(long id, long taskId, String notifyTime, int isSent) {
         this.id = id;
         this.taskId = taskId;
         this.notifyTime = notifyTime;
         this.isSent = isSent;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getTaskId() {
+    public long getTaskId() {
         return taskId;
     }
 
-    public void setTaskId(int taskId) {
+    public void setTaskId(long taskId) {
         this.taskId = taskId;
     }
